requirements.txt
#server
Django==4.0.3
psycopg2==2.9.1
Pillow==9.1.0
djangorestframework==3.13.1
drf-spectacular==0.20.2
web3==5.30.0

#heroku
gunicorn==20.1.0
